+++ R/reexport.R
@@ -71,4 +71,3 @@ gtsummary::style_ratio
 #' @export
 #' @importFrom gtsummary add_n
 gtsummary::add_n
-